requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
gpg==1.15.1
idna==3.7
libcomps==0.1.20
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rpm==4.16.1.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
